Update build.ps1 to check if files exists

+++ build.ps1
@@ -1,5 +1,8 @@
-Remove-Item -Path __pycache__ -Recurse
-Remove-Item -Path build -Recurse
-Remove-Item -Path dist -Recurse
-Remove-Item -Path main.spec
+Set-Alias exist Test-Path -Option "Constant, AllScope"
+
+if (exist __pycache) { Remove-Item -Path __pycache__ -Recurse }
+if (exist build) { Remove-Item -Path build -Recurse }
+if (exist dist) { Remove-Item -Path dist -Recurse }
+if (exist main.spec) { Remove-Item -Path main.spec }
+
 pyinstaller --onefile --hidden-import babel.numbers main.py
